Added files via upload
The sole purpose of this API is to protect your passwords and other important data from getting hacked. This API basically focus on the encryption of hardcoded credentials. It is well known fact that any .jar or .class file can be reverse engineered and any experienced can snoop your hard coded passwords. "Hard Coding" means something that you want to embed with your program or any project that cannot be changed directly. For example if you are using a database server, then you must hardcode the credentials to connect your database with your project and that cannot be changed by user. Because you have hard coded. Sometimes embedded data can be confidential and we would never want to reveal it to the user/snooper.so basically, hard coding is not a good practice it is one of the top coding mistakes made by programmers that lead their application to be hacked. There are several methods available on internet to avoid the hardcoding of the important data. But almost all the methods are very